chore: enforce 2000-line limit on source files (CI, ratcheting baseline)
- tools/check-file-lines.js: new files must stay <= 2000 lines;
  pre-existing offenders are grandfathered in file-lines-baseline.json
  and may only shrink. Generated/vendored code excluded (web/pkg,
  third_party, metal-patched fork, generated watchOS stubs).
- ffi_core.rs (2946 lines) split into ffi_core/<subsystem>.rs section
  macros composed by define_core_ffi! — nested $crate macro invocations
  keep call-site hygiene so the platform hooks still resolve. Largest
  section is 646 lines. Shared tests green, validator green, macOS
  release staticlib exports still byte-identical to the pre-macro
  baseline.
- remaining giants tracked for splitting: renderer/mod.rs (13956),
  renderer/shaders.rs (5861), web/src/lib.rs (2747),
  material_system.rs (2118), bloom-reference main.rs (2330)
