+++ tools/check-file-lines.js
@@ -1,0 +1,82 @@
+#!/usr/bin/env node
+// check-file-lines.js — enforce the 2000-line ceiling on source files.
+//
+// Rules:
+//   - Source files (.rs/.ts/.js/.swift/.wgsl) outside the exclusions must
+//     be <= LIMIT lines.
+//   - Files in tools/file-lines-baseline.json are grandfathered at their
+//     recorded size and may only shrink (ratchet). When one drops below
+//     LIMIT, remove it from the baseline.
+//   - Generated and vendored code is excluded — the limit exists for
+//     human-maintained files.
+//
+// Run: node tools/check-file-lines.js          (CI gates on this)
+//      node tools/check-file-lines.js --update (rewrite baseline after
+//                                               legitimately shrinking a file)
+'use strict';
+const fs = require('fs');
+const path = require('path');
+
+const ROOT = path.join(__dirname, '..');
+const LIMIT = 2000;
+const BASELINE_PATH = path.join(__dirname, 'file-lines-baseline.json');
+
+const INCLUDE_DIRS = ['src', 'native', 'tools', 'examples'];
+const EXTENSIONS = new Set(['.rs', '.ts', '.js', '.swift', '.wgsl']);
+const EXCLUDE = [
+  /node_modules\//,
+  /\/target\//,
+  /native\/third_party\//,        // vendored (JoltPhysics, shims)
+  /native\/web\/pkg\//,           // wasm-pack output, generated
+  /native\/tvos\/metal-patched\//,// vendored patched fork of metal-rs
+  /native\/watchos\/src\/ffi_stubs\.rs$/, // generated by gen_stubs.js
+];
+
+function* walk(dir) {
+  for (const e of fs.readdirSync(dir, { withFileTypes: true })) {
+    const p = path.join(dir, e.name);
+    if (e.isDirectory()) yield* walk(p);
+    else yield p;
+  }
+}
+
+const update = process.argv.includes('--update');
+const baseline = fs.existsSync(BASELINE_PATH)
+  ? JSON.parse(fs.readFileSync(BASELINE_PATH, 'utf8'))
+  : {};
+
+let failures = 0;
+const newBaseline = {};
+for (const dir of INCLUDE_DIRS) {
+  const abs = path.join(ROOT, dir);
+  if (!fs.existsSync(abs)) continue;
+  for (const file of walk(abs)) {
+    const rel = path.relative(ROOT, file);
+    if (!EXTENSIONS.has(path.extname(file))) continue;
+    if (EXCLUDE.some((re) => re.test(rel))) continue;
+    const lines = (fs.readFileSync(file, 'utf8').match(/\n/g) || []).length; // == wc -l
+    if (lines <= LIMIT) {
+      if (baseline[rel]) console.log(`note  ${rel} is now under ${LIMIT} lines — remove it from tools/file-lines-baseline.json`);
+      continue;
+    }
+    const allowed = baseline[rel];
+    if (allowed === undefined) {
+      console.error(`FAIL  ${rel}: ${lines} lines (limit ${LIMIT}). Split it — new files must stay under the limit.`);
+      failures++;
+    } else if (lines > allowed) {
+      console.error(`FAIL  ${rel}: ${lines} lines, grew past its grandfathered ${allowed}. Shrink it (or split it) — the baseline only ratchets down.`);
+      failures++;
+      newBaseline[rel] = allowed;
+    } else {
+      newBaseline[rel] = lines; // ratchet down to current size
+    }
+  }
+}
+
+if (update) {
+  fs.writeFileSync(BASELINE_PATH, JSON.stringify(newBaseline, null, 2) + '\n');
+  console.log(`baseline updated: ${Object.keys(newBaseline).length} grandfathered files`);
+}
+
+console.log(`${failures} failures`);
+process.exit(failures ? 1 : 0);
